Implement syscall handling.

+++ src/kernel/ata.h
@@ -5,5 +5,5 @@
 #define SECTOR_SIZE 512
 
 void ata_init();
-u8 *ata_read_sector(u32 offset);
+u8 *ata_read_sector(size_t offset);
 u8 *ata_read(size_t sector_offset, size_t length);
